Introduce pattern matching for conditions
Regex is too broad and introduces security issues. This pattern matching algorithm is based on Glob but without the file-specific `*` matching. Glob is not a standard and it therefore better to define our own pattern match. This is based on Proposal 4 of GA4GH Passport Visa Conditions. The goal is powerful enough pattern matching without too much complexity in implementation.